Add files via upload
Integrate a classifier, which can classify the voice commands, into a food ordering system. 
 **Objectives:**

Stage 0 - Obtain Features from Audio samples using Pre-trained Network

Stage 1- Train a classifier on Studio data and deploy the model in the server 

Stage 2 - Use 'Noisy_data' and 'Studio_data' together, train a classifier on the same, and deploy the model in the server.

Stage 3  - Collect your voice samples (team data) and refine the classifier trained on Studio_data and Noisy_data. Deploy the model in the server.